Debugged data_provider
