+++ web/views.py
@@ -7,7 +7,7 @@ from core.serializer import *
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import status
-
+import tempfile
 #from django.http import JsonResponse
 
 # Create your views here.
@@ -311,7 +311,7 @@ def page_report(request, report, fmt='csv', conf=None):
     fn = tf.name
     fp = open(fn, "w+")
     DataProvider.WriteData(objs, fmt, fn)
-    fclose(fp)
+    fp.close()
     response = StreamingHttpResponse(open(fn), content_type='text/csv') 
     response['Content-Disposition'] = 'attachment; filename=' + fn + ".csv"
     return response
